% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt.fixed_design}
\alias{as_gt.fixed_design}
\title{This is the function to format the bounds summary table of fixed design into gt style.}
\usage{
\method{as_gt}{fixed_design}(x, title = NULL, footnote = NULL, ...)
}
\arguments{
\item{x}{a summary object of group sequential design}

\item{title}{title to be displayed}

\item{footnote}{footnotes to be displayed}

\item{...}{additional arguments}
}
\value{
a gt table
}
\description{
This is the function to format the bounds summary table of fixed design into gt style.
}
\examples{
library(dplyr)
library(tibble)

# Enrollment rate
enroll_rate <- tibble(
  stratum = "All", 
  duration = 18, 
  rate = 20)

# Failure rates
fail_rate <- tibble(
  stratum = "All", 
  duration = c(4, 100), 
  fail_rate = log(2) / 12,
  hr = c(1, .6), 
  dropout_rate = .001)

# Study duration in months
study_duration <- 36

# Experimental / Control randomization ratio
ratio <- 1 

# 1-sided Type I error
alpha <- 0.025 

# Type II error (1 - power)
beta <- 0.1 

# ------------------------- #
#        AHR                #
# ------------------------- #
# under fixed power 
fixed_design(
  "ahr", 
  alpha = alpha, power = 1 - beta, 
  enroll_rate = enroll_rate, fail_rate = fail_rate, 
  study_duration = study_duration, ratio = ratio
  ) \%>\% 
  summary() \%>\% 
  as_gt()
  
# ------------------------- #
#        FH                 #
# ------------------------- #
# under fixed power
fixed_design(
  "fh", 
  alpha = alpha, power = 1 - beta, 
  enroll_rate = enroll_rate, fail_rate = fail_rate,
  study_duration = study_duration, ratio = ratio
  ) \%>\% 
  summary() \%>\% 
  as_gt()
  
}
