% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt}
\alias{as_gt}
\title{Deprecated: use lt() instead}
\usage{
as_gt(x, ...)
}
\arguments{
\item{x}{A summary object of a fixed or group sequential design.}

\item{...}{Additional arguments passed to \code{\link[lt:lt]{lt::lt()}}.}
}
\value{
An \code{lt_tbl} object.
}
\description{
\code{as_gt()} is deprecated; use \code{\link[lt:lt]{lt::lt()}} instead.
}
\seealso{
\code{\link[lt:lt]{lt::lt()}}, \link{lt-methods}
}
