% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rtf.R
\name{as_rtf}
\alias{as_rtf}
\alias{as_rtf.fixed_design}
\alias{as_rtf.gs_design}
\title{Convert summary table of a fixed or group sequential design object to a rtf object}
\usage{
as_rtf(x, ...)

\method{as_rtf}{fixed_design}(
  x,
  title = NULL,
  footnote = NULL,
  col_rel_width = NULL,
  orientation = "portrait",
  text_font_size = 9,
  path_outtable = NULL,
  ...
)

\method{as_rtf}{gs_design}(
  x,
  title = NULL,
  subtitle = NULL,
  colname_spanner = "Cumulative boundary crossing probability",
  colname_spannersub = c("Alternate hypothesis", "Null hypothesis"),
  footnote = NULL,
  display_bound = c("Efficacy", "Futility"),
  display_columns = NULL,
  display_inf_bound = TRUE,
  full_alpha = 0.025,
  col_rel_width = NULL,
  orientation = "portrait",
  text_font_size = 9,
  path_outtable = NULL,
  ...
)
}
\arguments{
\item{x}{A summary object of a fixed or group sequential design.}

\item{...}{}

\item{title}{A string to specify the title of the rtf table.}

\item{footnote}{A list containing \code{content}, \code{location}, and \code{attr}.
\code{content} is a vector of string to specify the footnote text;
\code{location} is a vector of string to specify the locations to put the
superscript of the footnote index;
\code{attr} is a vector of string to specify the attributes of the footnotes,
for example, \code{c("colname", "title", "subtitle", "analysis", "spanner")};
users can use the functions in the \code{gt} package to customize the table.}

\item{col_rel_width}{}

\item{orientation}{}

\item{text_font_size}{}

\item{path_outtable}{}

\item{subtitle}{A string to specify the subtitle of the rtf table.}

\item{colname_spanner}{A string to specify the spanner of the rtf table.}

\item{colname_spannersub}{A vector of strings to specify the spanner details of the rtf table.}

\item{display_bound}{A vector of strings specifying the label of the bounds.
The default is \code{c("Efficacy", "Futility")}.}

\item{display_columns}{A vector of strings specifying the variables to be
displayed in the summary table.}

\item{display_inf_bound}{Logical, whether to display the +/-inf bound.}

\item{full_alpha}{The full alpha used in the design, the default is 0.025.
when the futility bound is non-binding, a footnote will be displayed, saying
the smaller value subtracts the probability of crossing a futility bound before
crossing an efficacy bound at a later analysis under the null hypothesis.}
}
\value{
A \code{rtf} object.
}
\description{
Convert summary table of a fixed or group sequential design object to a rtf object
}
\examples{
library(dplyr)
library(tibble)

# Enrollment rate
enroll_rate <- define_enroll_rate(
  duration = 18,
  rate = 20
)

# Failure rates
fail_rate <- define_fail_rate(
  duration = c(4, 100),
  fail_rate = log(2) / 12,
  dropout_rate = .001,
  hr = c(1, .6)
)

# Study duration in months
study_duration <- 36

# Experimental / Control randomization ratio
ratio <- 1

# 1-sided Type I error
alpha <- 0.025

# Type II error (1 - power)
beta <- 0.1

# ------------------------- #
#        AHR                #
# ------------------------- #
# under fixed power
fixed_design_ahr(
  alpha = alpha, power = 1 - beta,
  enroll_rate = enroll_rate, fail_rate = fail_rate,
  study_duration = study_duration, ratio = ratio
) \%>\%
  summary() \%>\%
  as_rtf(path_outtable = tempfile(fileext = ".rtf"))

# ------------------------- #
#        FH                 #
# ------------------------- #
# under fixed power
fixed_design_fh(
  alpha = alpha, power = 1 - beta,
  enroll_rate = enroll_rate, fail_rate = fail_rate,
  study_duration = study_duration, ratio = ratio
) \%>\%
  summary() \%>\%
  as_rtf(path_outtable = tempfile(fileext = ".rtf"))
}
