% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg.R
\name{check_analysisTimes}
\alias{check_analysisTimes}
\title{A function to check the arguments \code{analysisTimes} used in gsDesign2}
\usage{
check_analysisTimes(analysisTimes)
}
\arguments{
\item{analysisTimes}{}
}
\value{
TURE or FALSE
}
\description{
A function to check the arguments \code{analysisTimes} used in gsDesign2
}
\examples{
analysisTimes <- 20
check_analysisTimes(analysisTimes)

analysisTimes <- c(20, 30)
check_analysisTimes(analysisTimes)
}
