% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg.R
\name{check_enrollRates}
\alias{check_enrollRates}
\title{A function to check the arguments \code{enrollRates} used in gsDesign2}
\usage{
check_enrollRates(enrollRates)
}
\arguments{
\item{enrollRates}{}
}
\value{
TURE or FALSE
}
\description{
A function to check the arguments \code{enrollRates} used in gsDesign2
}
\examples{

enrollRates <- tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6, 9))
check_enrollRates(enrollRates)

}
