% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg.R
\name{check_enrollRates_failRates}
\alias{check_enrollRates_failRates}
\title{A function to check the arguments \code{enrollRates} and \code{failRates} used in gsDesign2}
\usage{
check_enrollRates_failRates(enrollRates, failRates)
}
\arguments{
\item{enrollRates}{}
}
\value{
TURE or FALSE
}
\description{
A function to check the arguments \code{enrollRates} and \code{failRates} used in gsDesign2
}
\examples{

enrollRates <- tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6, 9))
failRates <- tibble::tibble(Stratum = "All", duration = c(3, 100), 
                            failRate = log(2) / c(9, 18), hr = c(.9, .6), 
                            dropoutRate = rep(.001, 2))
check_enrollRates(enrollRates, failRates)

}
