% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg.R
\name{check_events}
\alias{check_events}
\title{A function to check the arguments \code{events} used in gsDesign2}
\usage{
check_events(events)
}
\arguments{
\item{analysisTimes}{}
}
\value{
TURE or FALSE
}
\description{
A function to check the arguments \code{events} used in gsDesign2
}
\examples{
events <- 20
check_events(events)

events <- c(20, 30)
check_events(events)
}
