% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg.R
\name{check_ratio}
\alias{check_ratio}
\title{A function to check the arguments \code{ratio} used in gsDesign2}
\usage{
check_ratio(ratio)
}
\arguments{
\item{ratio}{}
}
\value{
TURE or FALSE
}
\description{
A function to check the arguments \code{ratio} used in gsDesign2
}
\examples{
ratio <- 1
check_ratio(ratio)

}
