% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg.R
\name{check_totalDuration}
\alias{check_totalDuration}
\title{A function to check the arguments \code{totalDuration} used in gsDesign2}
\usage{
check_totalDuration(totalDuration)
}
\arguments{
\item{totalDuration}{}
}
\value{
TURE or FALSE
}
\description{
A function to check the arguments \code{totalDuration} used in gsDesign2
}
\examples{
totalDuration <- 36
check_totalDuration(totalDuration)

totalDuration <- c(36, 48)
check_totalDuration(totalDuration)
}
