% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_enroll_rate}
\alias{define_enroll_rate}
\title{Define Enrollment Rate}
\usage{
define_enroll_rate(duration, rate)
}
\arguments{
\item{duration}{a numeric vector of piecewise study duration interval.}

\item{rate}{a numeric vector of enrollment rate in each \code{duration}.}
}
\description{
Define Enrollment Rate
}
\examples{

define_enroll_rate(
  duration = c(2, 2, 10),
  rate = c(3, 6, 9)
)

}
