% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eAccrual.R
\name{eAccrual}
\alias{eAccrual}
\title{Piecewise constant expected accrual}
\usage{
eAccrual(
  x = 0:24,
  enrollRates = tibble(duration = c(3, 3, 18), rate = c(5, 10, 20))
)
}
\arguments{
\item{x}{times at which enrollment is to be computed.}

\item{enrollRates}{Piecewise constant enrollment rates expressed as a \code{tibble} with
\code{duration} for each piecewise constant period and the \code{rate} of enrollment for that period.}
}
\value{
A vector with expected cumulative enrollment for the specified \code{times}.
}
\description{
\code{eAccrual()} computes the expected cumulative enrollment (accrual)
given a set of piecewise constant enrollment rates and times.
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Validate if input x is a vector of strictly increasing non-negative numeric elements.
   \item Validate if input enrollment rate is of type data.frame.
   \item Validate if input enrollment rate contains duration column.
   \item Validate if input enrollment rate contains rate column.
   \item Validate if rate in input enrollment rate is non-negative with at least one positive rate.
   \item Convert rates to step function.
   \item Add times where rates change to enrollment rates.
   \item Make a tibble of the input time points x, duration, enrollment rates at points, and
   expected accrual.
   \item Extract the expected cumulative or survival enrollment.
   \item Return \code{eAccrual}
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
library(tibble)

# Example 1: default
eAccrual()

# Example 2: unstratified design
eAccrual(x = c(5, 10, 20), 
         enrollRates = tibble(duration = c(3, 3, 18), rate = c(5, 10, 20)))

eAccrual(x = c(5, 10, 20), 
         enrollRates = tibble(duration = c(3, 3, 18), rate = c(5, 10, 20), 
         Stratum = "All"))
         
# Example 3: stratified design
eAccrual(x = c(24, 30, 40), 
         enrollRates = tibble(Stratum=c("subgroup", "complement"), 
                              duration = 33, 
                              rate = c(30, 30)))

}
