% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_design_combo.R
\name{get_combo_power}
\alias{get_combo_power}
\title{Function to calculate power}
\usage{
get_combo_power(n, bound, info_fh, theta_fh, corr_fh, algorithm, beta, ...)
}
\arguments{
\item{n}{Input sample size}

\item{algorithm}{ an object of class \code{\link[mvtnorm]{GenzBretz}},
                    \code{\link[mvtnorm]{Miwa}} or \code{\link[mvtnorm]{TVPACK}}
                    specifying both the algorithm to be used as well as
                    the associated hyper parameters.}

\item{beta}{Type II error.}

\item{...}{Additional parameters passed to \link[mvtnorm:pmvnorm]{mvtnorm::pmvnorm}.}
}
\value{
The optimal sample size (a single numeric value)
}
\description{
A helper function passed to \code{uniroot()}
}
\details{
This function calculates the difference between the derived power and the
targeted power (1 - beta), based on the provided sample size, upper and lower
boundaries, and treatment effect.
}
\keyword{internal}
