% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_bound_summary.R
\name{gs_bound_summary}
\alias{gs_bound_summary}
\title{Bound summary table}
\usage{
gs_bound_summary(x)
}
\arguments{
\item{x}{design object}
}
\value{
A data frame
}
\description{
Summarizes the efficacy and futility bounds for each analysis.
}
\examples{
library(gsDesign2)

x <- gs_design_ahr(info_frac = c(.25, .75, 1), analysis_time = c(12, 25, 36))
gs_bound_summary(x)

x <- gs_design_wlr(info_frac = c(.25, .75, 1), analysis_time = c(12, 25, 36))
gs_bound_summary(x)

}
\seealso{
\code{\link[gsDesign:gsBoundSummary]{gsDesign::gsBoundSummary()}}
}
