% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cp_npe.R
\name{gs_cp_npe}
\alias{gs_cp_npe}
\title{Conditional power computation with non-constant effect}
\usage{
gs_cp_npe(
  theta0 = NULL,
  theta1 = NULL,
  theta_hat = NULL,
  info0 = NULL,
  info1 = NULL,
  info_hat = NULL,
  info_scale = c("h0_h1_info", "h0_info", "h1_info"),
  i = NULL,
  j = NULL,
  z_i = NULL,
  z_j = NULL,
  local_alternative = TRUE
)
}
\arguments{
\item{theta0}{Natural parameter for null hypothesis (default is all zero).}

\item{theta1}{Natural parameter for alternate hypothesis.}

\item{theta_hat}{Natural parameter estimated based on observed blinded data.}

\item{info0}{Statistical information under null hypothesis.}

\item{info1}{Statistical information under alternative hypothesis.}

\item{info_hat}{Statistical information estimated based on observed blinded data.}

\item{info_scale}{Information scale for calculation. Options are:
\itemize{
\item \code{"h0_h1_info"} (default): variance under both null and alternative hypotheses is used.
\item \code{"h0_info"}: variance under null hypothesis is used.
\item \code{"h1_info"}: variance under alternative hypothesis is used.
}}

\item{i}{Index of current analysis.}

\item{j}{Index of the future analysis to calculate the conditional power.}

\item{z_i}{Interim z-value at analysis i (scalar).}

\item{z_j}{Efficacy boundary at analysis j (scalar).}

\item{local_alternative}{Logical value whether the local alternative hypothesis is used or not.}
}
\description{
Conditional power computation with non-constant effect
}
\examples{
library(gsDesign2)
gs_cp_npe(theta0 = c(0, 0, 0),
          theta1 = c(0.1, 0.2, 0.3),
          theta_hat = NULL,
          info0 = c(10, 20, 30),
          info1 = c(9.8, 19.6, 29.9),
          info_hat = NULL,
          info_scale = "h0_h1_info",
          i = 1, j = 3,
          z_i = 1.5, z_j = 1.96,
          local_alternative = TRUE)
}
