% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_info_combo.R
\name{gs_info_combo}
\alias{gs_info_combo}
\title{Information and effect size for max combo test}
\usage{
gs_info_combo(
  enroll_rate = tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6, 9)),
  fail_rate = tibble(Stratum = "All", duration = c(3, 100), fail_rate = log(2)/c(9, 18),
    hr = c(0.9, 0.6), dropout_rate = rep(0.001, 2)),
  ratio = 1,
  events = NULL,
  analysis_time = NULL,
  rho,
  gamma,
  tau = rep(-1, length(rho)),
  approx = "asymptotic"
)
}
\arguments{
\item{enroll_rate}{enrollment rates}

\item{fail_rate}{failure and dropout rates}

\item{ratio}{Experimental:Control randomization ratio (not yet implemented)}

\item{events}{Targeted events at each analysis}

\item{analysis_time}{Minimum time of analysis}

\item{rho}{Weighting parameters}

\item{gamma}{Weighting parameters}

\item{tau}{Weighting parameters}

\item{approx}{Approximation method}
}
\description{
Information and effect size for max combo test
}
