% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesign2-package.R
\docType{package}
\name{gsDesign2-package}
\alias{gsDesign2}
\alias{gsDesign2-package}
\title{gsDesign2: Group Sequential Design Under Non-Proportional Hazards}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compute sample size under non-proportional hazards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/gsDesign2/}
  \item \url{https://github.com/Merck/gsDesign2}
  \item Report bugs at \url{https://github.com/Merck/gsDesign2/issues}
}

}
\author{
\strong{Maintainer}: Yilong Zhang \email{yilong.zhang@merck.com}

Authors:
\itemize{
  \item Keaven Anderson \email{keaven_anderson@merck.com}
}

Other contributors:
\itemize{
  \item Amin Shirazi \email{ashirazist@gmail.com} [contributor]
  \item Ruixue Wang \email{ruixue.wang@merck.com} [contributor]
  \item Yi Cui \email{yi.cui@merck.com} [contributor]
  \item Ping Yang \email{ping.yang1@merck.com} [contributor]
  \item Xin Tong Li \email{xin.tong.li@merck.com} [contributor]
  \item Yalin Zhu \email{yalin.zhu@merck.com} [contributor]
  \item Nan Xiao \email{nan.xiao1@merck.com} [contributor]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
