% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_combo.R
\name{pmvnorm_combo}
\alias{pmvnorm_combo}
\title{Multivariate Normal Distribution for Multivariate Maximum Statistics}
\usage{
pmvnorm_combo(
  lower,
  upper,
  group,
  mean,
  corr,
  algorithm = GenzBretz(maxpts = 1e+05, abseps = 1e-05),
  ...
)
}
\arguments{
\item{lower}{ the vector of lower limits of length n.}

\item{upper}{ the vector of upper limits of length n.}

\item{group}{the vector of test statistics group.}

\item{mean}{ the mean vector of length n.}

\item{corr}{ the correlation matrix of dimension n.}

\item{algorithm}{ an object of class \code{\link[mvtnorm]{GenzBretz}},
                    \code{\link[mvtnorm]{Miwa}} or \code{\link[mvtnorm]{TVPACK}}
                    specifying both the algorithm to be used as well as
                    the associated hyper parameters.}

\item{...}{additional parameters transfer to \code{mvtnorm::pmvnorm}}
}
\description{
Computes the distribution function of the multivariate normal distribution
with maximum statistics for arbitrary limits and correlation matrices
}
\details{
Let $Z = {Z_ij}$ be a multivariate normal distribution.
Here i is a group indicator and j is a within group statistics indicator.
Let G_i = max({Z_ij}) for all test within one group.
This program are calculating the probability

$$Pr( lower < max(G) < upper )$$
}
