% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fixed_design}
\alias{summary.fixed_design}
\title{Summary for fixed design objects}
\usage{
\method{summary}{fixed_design}(object, ...)
}
\arguments{
\item{object}{A fixed design object returned by \code{\link[=fixed_design]{fixed_design()}}}

\item{...}{Additional arguments}
}
\value{
A data frame
}
\description{
Summary for \code{\link[=fixed_design]{fixed_design()}} objects
}
\examples{
library(dplyr)

# Enrollment rate
enroll_rate <- tibble::tibble(
  Stratum = "All",
  duration = 18,
  rate = 20)

# Failure rates
fail_rate <- tibble::tibble(
  Stratum = "All",
  duration = c(4, 100),
  fail_rate = log(2) / 12,
  hr = c(1, .6),
  dropout_rate = .001)

# Study duration in months
study_duration <- 36

# Experimental / Control randomization ratio
ratio <- 1

# 1-sided Type I error
alpha <- 0.025
# Type II error (1 - power)
beta <- 0.1

# ------------------------- #
#        AHR                #
# ------------------------- #
# under fixed power
fixed_design(
  "ahr",
  alpha = alpha,
  power = 1 - beta,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  study_duration = study_duration,
  ratio = ratio
  ) \%>\% summary()

# ------------------------- #
#        FH                 #
# ------------------------- #
# under fixed power
fixed_design(
  "fh",
  alpha = alpha,
  power = 1 - beta,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  study_duration = study_duration,
  ratio = ratio
  ) \%>\% summary()

}
