% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_integer.R
\name{to_integer.fixed_design}
\alias{to_integer.fixed_design}
\title{This is the function to format the fixed design into integer sample size.}
\usage{
\method{to_integer}{fixed_design}(x, sample_size = TRUE, ...)
}
\arguments{
\item{x}{an object returned from \code{fixed_design}}

\item{sample_size}{\code{TRUE} or \code{FALSE}, indicting if to ceiling sample size to an even integer}

\item{...}{additional arguments}
}
\value{
a list similar to the output of \code{fixed_design}, but the sample size is an integer
}
\description{
This is the function to format the fixed design into integer sample size.
}
\examples{
library(dplyr)
library(gsDesign2)

# Average hazard ratio
x <- fixed_design("ahr", 
                  alpha = .025, power = .9, 
                  enroll_rate = tibble::tibble(stratum = "All",  duration = 18, rate = 1),
                  fail_rate = tibble::tibble(stratum = "All", duration = c(4, 100), fail_rate = log(2) / 12, hr = c(1, .6), dropout_rate = .001),
                  study_duration = 36)
x \%>\% to_integer()    

# FH
x <- fixed_design("fh", alpha = 0.025, power = 0.9, 
                  enroll_rate = tibble::tibble(stratum = "All", duration = 18, rate = 20), 
                  fail_rate = tibble::tibble(stratum = "All", duration = c(4, 100), fail_rate = log(2) / 12,hr = c(1, .6), dropout_rate = .001), 
                  rho = 0.5, gamma = 0.5,
                  study_duration = 36, ratio = 1)         
x \%>\% to_integer() 

# MB
x <- fixed_design("mb", alpha = 0.025, power = 0.9, 
                  enroll_rate = tibble::tibble(stratum = "All", duration = 18, rate = 20), 
                  fail_rate = tibble::tibble(stratum = "All", duration = c(4, 100), fail_rate = log(2) / 12,hr = c(1, .6), dropout_rate = .001), 
                  tau = 4,
                  study_duration = 36, ratio = 1)         
x \%>\% to_integer()                   
}
