% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_integer.R
\name{to_integer.gs_design}
\alias{to_integer.gs_design}
\title{This is the function to format group sequential design into interger sample size.}
\usage{
\method{to_integer}{gs_design}(x, sample_size = TRUE, ...)
}
\arguments{
\item{x}{an object returned from \code{gs_design_ahr}, \code{gs_design_wlr}, or \code{gs_design_combo}}

\item{sample_size}{\code{TRUE} or \code{FALSE}, indicting if to ceiling sample size to an even integer}

\item{...}{additional arguments}
}
\value{
a list similar to the output of \code{gs_design_ahr}, \code{gs_design_wlr}, or \code{gs_design_combo},
but the sample size is an integer
}
\description{
This is the function to format group sequential design into interger sample size.
}
\examples{
library(dplyr)
gs_design_ahr() \%>\% to_integer()
gs_design_wlr() \%>\% to_integer()
}
