% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_integer.R
\name{to_integer}
\alias{to_integer}
\title{S3 class method to round sample size to an even number for equal design.}
\usage{
to_integer(x, ...)
}
\arguments{
\item{x}{an object returned from \code{gs_design_ahr}, \code{gs_design_wlr}, or \code{gs_design_combo}}

\item{...}{additional arguments}
}
\value{
a list similar to the output of \code{gs_design_ahr}, \code{gs_design_wlr}, or \code{gs_design_combo},
but the sample size is an integer
}
\description{
S3 class method to round sample size to an even number for equal design.
}
