% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ahr_blinded}
\alias{ahr_blinded}
\title{Blinded estimation of average hazard ratio}
\usage{
ahr_blinded(
  Srv = Surv(time = simtrial::Ex1delayedEffect$month, event =
    simtrial::Ex1delayedEffect$evntd),
  intervals = array(3, 3),
  hr = c(1, 0.6),
  ratio = 1
)
}
\arguments{
\item{Srv}{input survival object (see \code{Surv}); note that only 0=censored, 1=event for \code{Surv}}

\item{intervals}{Vector containing positive values indicating interval lengths where the
exponential rates are assumed.
Note that a final infinite interval is added if any events occur after the final interval
specified.}

\item{hr}{vector of hazard ratios assumed for each interval}

\item{ratio}{ratio of experimental to control randomization.}
}
\value{
A \code{tibble} with one row containing
\code{AHR} blinded average hazard ratio based on assumed period-specific hazard ratios input in \code{failRates}
and observed events in the corresponding intervals
\code{Events} total observed number of events, \code{info} statistical information based on Schoenfeld approximation,
and info0 (information under related null hypothesis) for each value of \code{totalDuration} input;
if \code{simple=FALSE}, \code{Stratum} and \code{t} (beginning of each constant HR period) are also returned
and \code{HR} is returned instead of \code{AHR}
}
\description{
Based on blinded data and assumed hazard ratios in different intervals, compute
a blinded estimate of average hazard ratio (AHR) and corresponding estimate of statistical information.
This function is intended for use in computing futility bounds based on spending assuming
the input hazard ratio (hr) values for intervals specified here.
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Validate if input hr is a numeric vector.
   \item Validate if input hr is non-negative.
   \item Simulate piece-wise exponential survival estimation with the inputs survival object Srv
   and intervals.
   \item Save the length of  hr and events to an object, and if the length of hr is shorter than
   the intervals, add replicates of the last element of hr and the corresponding numbers of events
   to hr.
   \item Compute the blinded estimation of average hazard ratio.
   \item Compute adjustment for information.
   \item Return a tibble of the sum of events, average hazard raito, blinded average hazard
   ratio, and the information.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
\dontrun{
library(simtrial)
library(survival)
ahr_blinded(Srv = Surv(time = simtrial::Ex2delayedEffect$month,
                       event = simtrial::Ex2delayedEffect$evntd),
            intervals = c(4, 100),
            hr = c(1, .55),
            ratio = 1)
}

}
