% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt}
\alias{as_gt}
\title{Deprecated: superseded by \code{\link[=as_lt]{as_lt()}}}
\usage{
as_gt(x, ...)
}
\arguments{
\item{x}{A summary object of a fixed or group sequential design.}

\item{...}{Additional arguments passed to \code{\link[=as_lt]{as_lt()}}.}
}
\value{
An \code{lt_tbl} object.
}
\description{
\code{as_gt()} is deprecated; use \code{\link[=as_lt]{as_lt()}} instead. It now calls \code{\link[=as_lt]{as_lt()}}.
}
\seealso{
\code{\link[=as_lt]{as_lt()}}
}
