% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt}
\alias{as_gt}
\title{S3 class method to get summary table into a gt table}
\usage{
as_gt(x, ...)
}
\arguments{
\item{x}{a summary of fixed or group sequential design}

\item{...}{additional arguments}
}
\value{
a gt table
}
\description{
S3 class method to get summary table into a gt table
}
