% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_fail_rate}
\alias{define_fail_rate}
\title{Define fail rate}
\usage{
define_fail_rate(duration, fail_rate, dropout_rate, hr = 1, stratum = "All")
}
\arguments{
\item{duration}{A numeric vector of piecewise study duration interval.}

\item{fail_rate}{A numeric vector of failure rate in each \code{duration} in the control group.}

\item{dropout_rate}{A numeric vector of dropout rate in each \code{duration}.}

\item{hr}{A numeric vector of hazard ratio between case and control group.}

\item{stratum}{A character vector of stratum name.}
}
\value{
A \code{fail_rate} data frame.
}
\description{
The function define subject failure rate for a study with two treatment groups.
It also supports stratified design that have different failure rate in
each stratum.
}
\examples{
# Define enroll rate
define_fail_rate(
  duration = c(3, 100),
  fail_rate = log(2) / c(9, 18),
  hr = c(.9, .6),
  dropout_rate = .001
)

# Define enroll rate with stratum
define_fail_rate(
  stratum = c(rep("Low", 2), rep("High", 2)),
  duration = 1,
  fail_rate = c(.1, .2, .3, .4),
  dropout_rate = .001,
  hr = c(.9, .75, .8, .6)
)
}
