% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_design.R
\name{fixed_design}
\alias{fixed_design}
\title{Fixed design sample size}
\usage{
fixed_design(
  x = c("AHR", "FH", "MB", "LF", "RD", "MaxCombo", "RMST", "Milestone"),
  alpha = 0.025,
  power = NULL,
  ratio = 1,
  studyDuration = 36,
  ...
)
}
\arguments{
\item{x}{Sample size method; default is \code{"AHR"};
other options include \code{"FH"}, \code{"MB"}, \code{"LF"}, \code{"RD"}, \code{"MaxCombo"}, \code{"Milestone"}.}

\item{alpha}{One-sided Type I error (strictly between 0 and 1)}

\item{power}{Power (\code{NULL} to compute power or strictly between 0 and \code{1 - alpha} otherwise)}

\item{ratio}{Experimental:Control randomization ratio}

\item{studyDuration}{study duration}

\item{...}{additional arguments like \code{enrollRates}, \code{failRates}, \code{rho}, \code{gamma}, \code{tau}}
}
\value{
a table
}
\description{
Computes fixed design sample size for many sample size methods.
Returns a \code{tibble} with a basic summary
}
\examples{
library(dplyr)

# Average hazard ratio
x <- fixed_design("AHR", 
                  alpha = .025, power = .9, 
                  enrollRates = tibble::tibble(Stratum = "All",  duration = 18, rate = 1),
                  failRates = tibble::tibble(Stratum = "All", duration = c(4, 100), failRate = log(2) / 12, hr = c(1, .6), dropoutRate = .001),
                  studyDuration = 36)
x \%>\% summary()            

# Lachin and Foulkes (uses gsDesign::nSurv())
x <- fixed_design("LF", 
                  alpha = .025, power = .9, 
                  enrollRates = tibble::tibble(Stratum = "All",  duration = 18, rate = 1),
                  failRates = tibble::tibble(Stratum = "All", duration = 100, failRate = log(2) / 12, hr = .7, dropoutRate = .001),
                  studyDuration = 36)
x \%>\% summary()

# RMST
x <- fixed_design("RMST", alpha = .025, power = .9, 
                  enrollRates = tibble::tibble(Stratum = "All",  duration = 18, rate = 1),
                  failRates = tibble::tibble(Stratum = "All", duration = 100, failRate = log(2) / 12, hr = .7, dropoutRate = .001),
                  studyDuration = 36,
                  tau = 18)
x \%>\% summary()

# Milestone 
x <- fixed_design("Milestone", alpha = .025, power = .9, 
                  enrollRates = tibble::tibble(Stratum = "All",  duration = 18, rate = 1),
                  failRates = tibble::tibble(Stratum = "All", duration = 100, failRate = log(2) / 12, hr = .7, dropoutRate = .001),
                  studyDuration = 36,
                  tau = 18)
x \%>\% summary()

}
