% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cp.R
\name{gs_cp}
\alias{gs_cp}
\title{Conditional power computation with non-constant effect size for non-/crossing an upper boundary at analysis j given observed Z value at analysis i}
\usage{
gs_cp(x = NULL, theta = NULL, i = 1, zi = NULL)
}
\arguments{
\item{x}{An object of type gsDesign2.}

\item{theta}{Optional numeric vector with length \eqn{j-i+1}, which specifies
the natural parameter for treatment effect of interim analysis \eqn{i} through
analysis \eqn{j}. The default is \code{NULL}.}

\item{i}{Index of current analysis, with default of 1.}

\item{zi}{Numeric scalar z-value observed at analysis \eqn{i}.}
}
\value{
A list of conditional powers:
\itemize{
\item \code{prob_alpha} is a numeric vector of
(\eqn{\alpha_{i,i+1}, ..., \alpha_{i,j-1}, \alpha_{i,j}}), where
\itemize{
\item for \eqn{j = i+1},
\eqn{\alpha_{i,j} = P(\{Z_j \geq b_j\} \mid Z_i = z_i)};
\item for \eqn{j > i+1},
\eqn{\alpha_{i,j} = P(\{Z_j \geq b_j\} \& \{\cap_{m=i+1}^{j-1} a_m \leq Z_m < b_m\} \mid Z_i = z_i)}.
}
\item \code{prob_alpha_plus} is a numeric vector of
(\eqn{\alpha^+_{i,i+1}, ..., \alpha^+_{i,j-1}, \alpha^+_{i,j}}), where
\itemize{
\item for \eqn{j = i+1},
\eqn{\alpha^+_{i,j} = P(\{Z_j \geq b_j\} \mid Z_i = z_i)};
\item for \eqn{j > i+1},
\eqn{\alpha^+_{i,j} = P(\{Z_j \geq b_j\} \& \{\cap_{m=i+1}^{j-1} Z_m < b_m\} \mid Z_i = z_i)}.
}
\item \code{prob_beta} is a numeric vector of
(\eqn{\beta_{i,i+1}, ..., \beta_{i,j-1}, \beta_{i,j}}), where
\itemize{
\item for \eqn{j = i+1},
\eqn{\beta_{i,j} = P(\{Z_j \leq a_j\} \mid Z_i = z_i)};
\item for \eqn{j > i+1},
\eqn{\beta_{i,j} = P(\{Z_j \leq a_j\} \& \{\cap_{m=i+1}^{j-1} a_m \leq Z_m < b_m\} \mid Z_i = z_i)}.
}
}
}
\description{
Conditional power computation with non-constant effect size for non-/crossing an upper boundary at analysis j given observed Z value at analysis i
}
\examples{
library(gsDesign2)
library(gsDesign)
library(dplyr)

enroll_rate <- define_enroll_rate(duration = c(2, 2, 2, 18),
                                  rate = c(1, 2, 3, 4))
fail_rate <- define_fail_rate(duration = c(3, Inf),
                              fail_rate = log(2) / 10,
                              dropout_rate = 0.001,
                              hr = c(1, 0.7))

# Example 1: futility is tested at all analyses
# original design ----
x <- gs_design_ahr(enroll_rate = enroll_rate, fail_rate = fail_rate,
                   alpha = 0.025, beta = 0.1, ratio = 1,
                   info_frac = c(0.4, 0.6, 0.8, 1), analysis_time = 30,
                   binding = FALSE,
                   upper = gs_spending_bound,
                   upar = list(sf = sfLDOF, total_spend = 0.025, param = NULL),
                   lower = gs_spending_bound,
                   lpar = list(sf = sfLDOF, total_spend = 0.1),
                   h1_spending = TRUE,
                   test_lower = TRUE,
                   info_scale = "h0_h1_info") |> to_integer()

# calculate conditional power
# case 1: currently at IA1, compute conditional power at IA2, IA3 and FA, 
# with default theta = NULL
gs_cp(x = x, i = 1, 
zi = -gsDesign::hrn2z(hr = 0.8, n = 150+180, ratio = 1))

# case 2: currently at IA1, compute conditional power at IA2, IA3 and FA, 
# with user-input theta
gs_cp(x = x, 
  theta = c(0.15, 0.2, 0.25, 0.3), 
  i = 1, 
  zi = -gsDesign::hrn2z(hr = 0.8, n = 150+180, ratio = 1))

# Example 2: futility is only tested at IA1
# original design ----
x <- gs_design_ahr(enroll_rate = enroll_rate, fail_rate = fail_rate,
                   alpha = 0.025, beta = 0.1, ratio = 1,
                   info_frac = c(0.4, 0.6, 0.8, 1), analysis_time = 30,
                   binding = FALSE,
                   upper = gs_spending_bound,
                   upar = list(sf = sfLDOF, total_spend = 0.025, param = NULL),
                   test_upper = c(FALSE, TRUE, TRUE, TRUE),
                   lower = gs_spending_bound,
                   lpar = list(sf = sfLDOF, total_spend = 0.1),
                   test_lower = c(TRUE, FALSE, FALSE, FALSE),
                   h1_spending = TRUE,
                   info_scale = "h0_h1_info") |> to_integer()

# calculate conditional power
# case 1: currently at IA1, compute conditional power at IA2, IA3 and FA,
# with default theta = NULL
gs_cp(x = x, i = 1,
      zi = -gsDesign::hrn2z(hr = 0.8, n = 150+180, ratio = 1))

# case 2: currently at IA1, compute conditional power at IA2, IA3 and FA,
# with user-input theta
gs_cp(x = x,
  theta = c(0.15, 0.2, 0.25, 0.3),
  i = 1,
  zi = -gsDesign::hrn2z(hr = 0.8, n = 150+180, ratio = 1))
}
