% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cp.R
\name{gs_cp}
\alias{gs_cp}
\title{Conditional probability under non-proportional hazards for
group sequential design}
\usage{
gs_cp(x, i, zi, j)
}
\arguments{
\item{x}{An object of type \code{gs_design}.}

\item{i}{Analysis at which interim z-value is given; must be from 1 to
\code{max(x$analysis$analysis)}.}

\item{zi}{Interim z-value at analysis \code{i} (scalar).}
}
\value{
A list containing:
\itemize{
\item \code{theta} - A list containing \code{theta0} (\code{theta} under H0, i.e., 0)
and \code{theta1} (a vector of \code{theta} under H1).
\item \code{upper_bound} - The upper bound value return by any
\code{gs_design_ahr()} functions.
\item \code{upper_prob} - A list containing the conditional probability given
\code{zi} under H0 and H1, respectively.
}
}
\description{
Computes conditional boundary crossing probabilities at future planned
analyses for a given group sequential design assuming an interim
z-statistic at a specified interim analysis.
}
\examples{
x <- gs_design_ahr(info_frac = c(.25, .75, 1), analysis_time = 36)
gs_cp(x = x, i = 1, zi = 0, j = 2)
}
