% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cp_ahr.R
\name{gs_cp_ahr}
\alias{gs_cp_ahr}
\title{Conditional power computation for group sequential designs using the average hazard ratio method.}
\usage{
gs_cp_ahr(x = NULL, i = 1, z_i = 0, hr_future = NULL)
}
\arguments{
\item{x}{A design object of type \code{gs_design_ahr()},  \code{gs_power_ahr()}, or \code{gs_update_ahr()}.}

\item{i}{Analysis at which interim z-value is given; must be from 1 to \code{nrow(x$analysis) - 1}.}

\item{z_i}{Interim z-value at analysis \code{i} (scalar).}

\item{hr_future}{A scalar or vector, which specifies the natural parameter for treatment effect at the future analyses.
The default gives you a range of possible HRs from 0.6 to 1.}
}
\value{
A data frame showing the conditional power based on the input design, interim test statistic, and possible future HR.
}
\description{
Conditional power computation for group sequential designs using the average hazard ratio method.
}
\examples{
library(gsDesign)
library(gsDesign2)

# Original design
x <- gs_design_ahr(info_frac = c(0.5, 0.8, 1))

# Updated design
xu <- gs_update_ahr(
  x = x,
  ustime = c(170, x$analysis$event[2:3] )/ max(x$analysis$event),
  event_tbl = data.frame(analysis = c(1, 1), event = c(100, 70)))

# We assume IA1 p-value of 0.04, one-sided, and future HR is possible from 0.6 to 1.
gs_cp_ahr(x = xu,
          i = 1,
          z_i = -qnorm(0.04),
          hr_future = seq(0.6, 1, by = 0.1))

# We assume the test statistics z_i is derived by the function hrn2z(), which translates
# a HR and number of events into an approximate corresponding Z-value,
# using the Schoenfeld approximation.
# The future HR is possible from 0.6 to 1.
gs_cp_ahr(x = xu,
          i = 1,
          z_i = -hrn2z(hr = 0.7, n = xu$analysis$event[1], ratio = 1),
          hr_future = seq(0.6, 1, by = 0.1))

}
