% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cp_simple.R
\name{gs_cp_simple}
\alias{gs_cp_simple}
\title{Simple conditional power computation with non-constant effect size}
\usage{
gs_cp_simple(x = NULL, theta = NULL, i = 1, zi = NULL)
}
\arguments{
\item{x}{An object of type gsDesign2.}

\item{theta}{Optional numeric vector of natural parameter for treatment effects from analysis \eqn{i} through final analysis \eqn{K}.
If \code{NULL}, values are taken from \code{x$analysis$theta}.}

\item{i}{Integer index for current analysis. Default is 1.}

\item{zi}{Numeric scalar z-value observed at analysis \eqn{i}.}
}
\value{
A numeric vector with the conditional power
\eqn{P(Z_j > b_j \mid Z_i = z_i)} for \eqn{j = i+1, ..., K}.
}
\description{
Simple conditional power computation with non-constant effect size
}
\details{
Suppose there are \eqn{K} analyses. Let \eqn{Z_i} and \eqn{Z_j} be the Z-statistics at a current analysis \eqn{i} and a future analysis \eqn{j}, respectively.
Let's denote the statistical information at the \eqn{i}-th analysis as \eqn{I_i}.
We further assume \eqn{Z_i} and \eqn{Z_j} are bivariate normal under standard group sequential assumptions with independent increments, then
\deqn{E(Z_i) = \theta_i\sqrt{I_i}}
\deqn{Var(Z_i) = 1/I_i}
\deqn{Cov(Z_i, Z_j) = t \equiv I_i/I_j}
See https://merck.github.io/gsDesign2/articles/story-npe-background.html for assumption details.
Returned value is
\deqn{P(Z_j > b_j \mid Z_i = z_i) = 1 - \Phi\left(\frac{b_j - \sqrt{t}z_i - \sqrt{I_j}(\theta_j - \theta_i\sqrt{t})}{\sqrt{1 - t}}\right)}
where \eqn{b_j} is the efficacy bound at analysis \eqn{j}, \eqn{z_i} is the observed Z-value at analysis \eqn{i},
and \eqn{\theta_i} is the natural parameter for treatment effect at analysis \eqn{i}.
}
\examples{
library(gsDesign2)
library(gsDesign)
# Calculate conditional power with optional user-input theta 
# (if NULL, will come from the input design)
alpha <- 0.025
beta <- 0.1
ratio <- 1
enroll_rate <- define_enroll_rate(
  duration = c(2, 2, 10),
  rate = (1:3) / 3)
fail_rate <- define_fail_rate(
  duration = Inf, fail_rate = log(2) / 9,
  hr = 0.6, dropout_rate = .0001)
analysis_time <- c(12, 24, 36)
ratio <- 1
upper <- gs_spending_bound
lower <- gs_b
upar <- list(sf = sfLDOF, total_spend = alpha)
lpar <- rep(-Inf, 3)
x <- gs_design_ahr(
  enroll_rate = enroll_rate, fail_rate = fail_rate,
  alpha = alpha, beta = beta, ratio = ratio,
  info_scale = "h0_h1_info",
  info_frac = NULL,
  analysis_time = c(12, 24, 36),
  upper = upper, upar = upar, test_upper = TRUE,
  lower = lower, lpar = lpar, test_lower = FALSE,
) |>
  to_integer()

# theta is NULL case
gs_cp_simple(x = x, theta = NULL, i = 1, zi = 1.5)

# User-input theta case
gs_cp_simple(x = x, theta = c(0.1, 0.2, 0.3), i = 1, zi = 1.5)

}
