% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_design_rd.R
\name{gs_design_rd}
\alias{gs_design_rd}
\title{Group sequential design using average hazard ratio under non-proportional hazards}
\usage{
gs_design_rd(
  p_c = tibble(Stratum = "All", Rate = 0.2),
  p_e = tibble(Stratum = "All", Rate = 0.15),
  IF = 1:3/3,
  rd0 = 0,
  alpha = 0.025,
  beta = 0.1,
  ratio = 1,
  stratum_prev = NULL,
  weight = c("un-stratified", "ss", "invar"),
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(0.1), rep(-Inf, 2)),
  test_upper = TRUE,
  test_lower = TRUE,
  info_scale = c(0, 1, 2),
  binding = FALSE,
  r = 18,
  tol = 1e-06,
  h1_spending = FALSE
)
}
\arguments{
\item{p_c}{rate at the control group}

\item{p_e}{rate at the experimental group}

\item{IF}{statistical information fraction}

\item{rd0}{treatment effect under super-superiority designs, the default is 0}

\item{alpha}{One-sided Type I error}

\item{beta}{Type II error}

\item{ratio}{Experimental:Control randomization ratio (not yet implemented)}

\item{stratum_prev}{randomization ratio of different stratum.
If it is un-stratified design then \code{NULL}.
Otherwise it is a tibble containing two columns (Stratum and prevalence).}

\item{weight}{the weighting scheme for stratified population}

\item{upper}{Function to compute upper bound}

\item{lower}{Function to compute lower bound}

\item{upar}{Parameter passed to \code{upper()}}

\item{lpar}{Parameter passed to \code{lower()}}

\item{test_upper}{indicator of which analyses should include an upper (efficacy) bound; single value of TRUE (default) indicates all analyses;
otherwise, a logical vector of the same length as \code{info} should indicate which analyses will have an efficacy bound}

\item{test_lower}{indicator of which analyses should include an lower bound; single value of TRUE (default) indicates all analyses;
single value FALSE indicated no lower bound; otherwise, a logical vector of the same length as \code{info} should indicate which analyses will have a
lower bound}

\item{info_scale}{the information scale for calculation}

\item{binding}{indicator of whether futility bound is binding; default of FALSE is recommended}

\item{r}{Integer, at least 2; default of 18 recommended by Jennison and Turnbull}

\item{tol}{Tolerance parameter for boundary convergence (on Z-scale)}

\item{h1_spending}{Indicator that lower bound to be set by spending under alternate hypothesis (input \code{failRates})
if spending is used for lower bound}
}
\value{
a \code{tibble} with columns Analysis, Bound, Z, Probability, theta, Time, AHR, Events
}
\description{
Group sequential design using average hazard ratio under non-proportional hazards
}
\details{
Need to be added
}
\examples{
library(tibble)
library(gsDesign)

# ----------------- #
#    example 1      #
#------------------ #
# un-stratified group sequential design
gs_design_rd(
  p_c = tibble(Stratum = "All", Rate = .2),
  p_e = tibble(Stratum = "All", Rate = .15),
  IF = c(0.7, 1),
  rd0 = 0, 
  alpha = .025,                  
  beta = .1,                    
  ratio = 1,
  stratum_prev = NULL,
  weight = "un-stratified",
  upper = gs_b,
  lower = gs_b,
  upar = gsDesign(k = 3, test.type = 1, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lpar = c(qnorm(.1), rep(-Inf, 2))
  )
  
# ----------------- #
#     example 2     #
# ----------------- #
# stratified group sequential design
gs_design_rd(
  p_c = tibble(Stratum = c("biomarker positive", "biomarker negative"), Rate = c(.2, .25)),
  p_e = tibble(Stratum = c("biomarker positive", "biomarker negative"), Rate = c(.15,.22)),
  IF = c(0.7, 1),
  rd0 = 0, 
  alpha = .025,                  
  beta = .1,                    
  ratio = 1,
  stratum_prev = tibble(Stratum = c("biomarker positive", "biomarker negative"), prevalence = c(.4, .6)),
  weight = "ss",
  upper = gs_spending_bound,lower = gs_b,
  upar = list(sf = gsDesign::sfLDOF, total_spend = 0.025, param = NULL, timing = NULL),
  lpar = rep(-Inf, 2)
)

}
