% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_info_ahr.R
\name{gs_info_ahr}
\alias{gs_info_ahr}
\title{Information and effect size based on AHR approximation}
\usage{
gs_info_ahr(
  enrollRates = tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6,
    9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  ratio = 1,
  events = NULL,
  analysisTimes = NULL
)
}
\arguments{
\item{enrollRates}{enrollment rates}

\item{failRates}{failure and dropout rates}

\item{ratio}{Experimental:Control randomization ratio}

\item{events}{Targeted minimum events at each analysis}

\item{analysisTimes}{Targeted minimum study duration at each analysis}
}
\value{
a \code{tibble} with columns \code{Analysis, Time, AHR, Events, theta, info, info0.}
\code{info, info0} contains statistical information under H1, H0, respectively.
For analysis \code{k}, \code{Time[k]} is the maximum of \code{analysisTimes[k]} and the expected time
required to accrue the targeted \code{events[k]}.
\code{AHR} is expected average hazard ratio at each analysis.
}
\description{
Based on piecewise enrollment rate, failure rate, and dropout rates computes
approximate information and effect size using an average hazard ratio model.
}
\details{
The \code{AHR()} function computes statistical information at targeted event times.
The \code{tEvents()} function is used to get events and average HR at targeted \code{analysisTimes}.
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Validate if input events is a numeric value vector or a vector with increasing values.
   \item Validate if input analysisTime is a numeric value vector or a vector with increasing values.
   \item Validate if inputs events and analysisTime have the same length if they are both specified.
   \item Compute average hazard ratio:
   \itemize{
     \item If analysisTime is specified, calculate average hazard ratio using \code{gsDesign2::AHR()}.
     \item If events is specified, calculate average hazard ratio using \code{gsDesign2::tEvents()}.
   }
   \item Return a tibble of Analysis, Time, AHR, Events, theta, info, info0.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
library(gsDesign)
library(gsDesign2)

# ------------------------ #
#       Example 1          #
# ------------------------ #
# Only put in targeted events
gs_info_ahr(events = c(30, 40, 50))

# ------------------------ #
#       Example 2          #
# ------------------------ #
# Only put in targeted analysis times
gs_info_ahr(analysisTimes = c(18, 27, 36))

# ------------------------ #
#       Example 3          #
# ------------------------ #
# Some analysis times after time at which targeted events accrue
# Check that both Time >= input analysisTime and Events >= input events
gs_info_ahr(events = c(30, 40, 50), analysisTimes = c(16, 19, 26))
gs_info_ahr(events = c(30, 40, 50), analysisTimes = c(14, 20, 24))

}
