% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_spending_combo.R
\name{gs_spending_combo}
\alias{gs_spending_combo}
\title{Derive spending bound for MaxCombo group sequential boundary}
\usage{
gs_spending_combo(par = NULL, info = NULL, ...)
}
\arguments{
\item{par}{a list with the following items:
\code{sf} (class spending function),
\code{total_spend} (total spend),
\code{param} (any parameters needed by the spending function \code{sf()}),
\code{timing} (a vector containing values at which spending function is to be evaluated or NULL if information-based spending is used),
\code{max_info} (when \code{timing} is NULL, this can be input as positive number to be used with \code{info} for information fraction at each analysis)}

\item{info}{statistical information at all analyses, at least up to analysis k}

\item{...}{additional parameters transfered to \code{par$sf}.}
}
\description{
Derive spending bound for MaxCombo group sequential boundary
}
\examples{

# alpha-spending
par <- list(sf = gsDesign::sfLDOF, total_spend = 0.025)
gs_spending_combo(par, info = 1:3/3)

# beta-spending
par <- list(sf = gsDesign::sfLDOF, total_spend = 0.2)
gs_spending_combo(par, info = 1:3/3)

}
