% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesign2.R
\docType{package}
\name{gsDesign2-package}
\alias{gsDesign2}
\alias{gsDesign2-package}
\title{gsDesign2: Group sequential design with non-constant effect}
\description{
Basic group sequential design computations extended.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LittleBeannie/gsDesign2}
  \item \url{https://littlebeannie.github.io/gsDesign2/}
  \item Report bugs at \url{https://github.com/LittleBeannie/gsDesign2/issues}
}

}
\author{
\strong{Maintainer}: Keaven Anderson \email{keaven_anderson@merck.com}

Authors:
\itemize{
  \item Yujie Zhao \email{yujie.zhao@merck.com}
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
