% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesign2-package.R
\docType{package}
\name{gsDesign2-package}
\alias{gsDesign2}
\alias{gsDesign2-package}
\title{gsDesign2: Group Sequential Design Under Non-Proportional Hazards}
\description{
Compute sample size under non-proportional hazards.
}
\author{
\strong{Maintainer}: Keaven Anderson \email{keaven_anderson@merck.com}

Authors:
\itemize{
  \item Yilong Zhang \email{yilong.zhang@merck.com}
}

Other contributors:
\itemize{
  \item Amin Shirazi \email{ashirazist@gmail.com} [contributor]
  \item Ruixue Wang \email{ruixue.wang@merck.com} [contributor]
  \item Yi Cui \email{yi.cui@merck.com} [contributor]
  \item Ping Yang \email{ping.yang1@merck.com} [contributor]
  \item Xin Tong Li \email{xin.tong.li@merck.com} [contributor]
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
