% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesign2-package.R
\docType{package}
\name{gsDesign2-package}
\alias{gsDesign2}
\alias{gsDesign2-package}
\title{gsDesign2: Group Sequential Design with Non-Constant Effect}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'gsDesign2' is to enable fixed or group sequential design under non-proportional hazards. To enable highly flexible enrollment, time-to-event and time-to-dropout assumptions, 'gsDesign2' offers piecewise constant enrollment, failure rates, and dropout rates for a stratified population. This package includes three methods for designs: average hazard ratio, weighted logrank tests in Yung and Liu (2019) \doi{10.1111/biom.13196}, and MaxCombo tests. Substantial flexibility on top of what is in the 'gsDesign' package is intended for selecting boundaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/gsDesign2/}
  \item \url{https://github.com/Merck/gsDesign2}
  \item Report bugs at \url{https://github.com/Merck/gsDesign2/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Authors:
\itemize{
  \item Keaven Anderson \email{keaven_anderson@merck.com}
  \item Yilong Zhang \email{elong0527@gmail.com}
  \item Jianxiao Yang \email{yangjx@ucla.edu}
  \item Nan Xiao \email{nan.xiao1@merck.com}
}

Other contributors:
\itemize{
  \item Amin Shirazi \email{ashirazist@gmail.com} [contributor]
  \item Ruixue Wang \email{ruixue.wang@merck.com} [contributor]
  \item Yi Cui \email{yi.cui@merck.com} [contributor]
  \item Ping Yang \email{ping.yang1@merck.com} [contributor]
  \item Xin Tong Li \email{xin.tong.li@merck.com} [contributor]
  \item Chenxiang Li \email{chenxiang.li@merck.com} [contributor]
  \item Hongtao Zhang \email{hongtao.zhang1@merck.com} [contributor]
  \item Yalin Zhu \email{yalin.zhu@merck.com} [contributor]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
