% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fixed_design}
\alias{summary.fixed_design}
\title{summary for \code{fixed_design()} object}
\usage{
\method{summary}{fixed_design}(x, ...)
}
\arguments{
\item{x}{a fixed design object returned by \code{fixed_design()}}

\item{...}{additional arguments}
}
\description{
summary function to fixed_design class
}
\examples{
library(dplyr)

# Enrollment rate
enrollRates <- tibble::tibble(
  Stratum = "All", 
  duration = 18, 
  rate = 20)

# Failure rates
failRates <- tibble::tibble(
  Stratum = "All", 
  duration = c(4, 100), 
  failRate = log(2) / 12,
  hr = c(1, .6), 
  dropoutRate = .001)

# Study duration in months
studyDuration <- 36

# Experimental / Control randomization ratio
ratio <- 1 

# 1-sided Type I error
alpha <- 0.025 
# Type II error (1 - power)
beta <- 0.1 

# ------------------------- #
#        AHR                #
# ------------------------- #
# under fixed power 
fixed_design(
  x = "AHR", 
  alpha = alpha, 
  power = 1 - beta, 
  enrollRates = enrollRates, 
  failRates = failRates, 
  studyDuration = studyDuration, 
  ratio = ratio
  ) \%>\% summary() 

# ------------------------- #
#        FH                 #
# ------------------------- #
# under fixed power
fixed_design(
  x = "FH", 
  alpha = alpha, 
  power = 1 - beta, 
  enrollRates = enrollRates, 
  failRates = failRates, 
  studyDuration = studyDuration,
   ratio = ratio
  ) \%>\% summary() 
  
}
