% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{S3 class method to summary fixed or group sequential design}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{a fixed design object or a group sequential design object}

\item{...}{additional arguments}
}
\value{
a R data frame
}
\description{
S3 class method to summary fixed or group sequential design
}
