% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr_weight.R
\name{wlr_weight}
\alias{wlr_weight}
\alias{wlr_weight_fh}
\alias{wlr_weight_1}
\alias{wlr_weight_n}
\title{Weight Function of Weighted Log-rank Test}
\usage{
wlr_weight_fh(x, arm0, arm1, rho = 0, gamma = 0, tau = NULL)

wlr_weight_1(x, arm0, arm1)

wlr_weight_n(x, arm0, arm1, power = 1)
}
\arguments{
\item{x}{analysis time}

\item{arm0}{an "arm" object defined in \code{npsurvSS} package}

\item{arm1}{an "arm" object defined in \code{npsurvSS} package}

\item{rho}{A scalar parameter that controls the type of test}

\item{gamma}{A scalar parameter that controls the type of test}

\item{tau}{A scalar parameter of the cut-off time for modest weighted log rank test}

\item{power}{A scalar parameter that controls the power of the weight function}
}
\description{
\itemize{
\item \code{wlr_weight_fh} is Fleming-Harriongton, FH(rho, gamma) weight function.
\item \code{wlr_weight_1}  is constant for log rank test
\item \code{wlr_weight_power} is Gehan-Breslow and Tarone-Ware weight function.
}
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Compute the sample size via the sum of arm sizes.
   \item Compute the proportion of size in the two arms.
   \item If the input tau is specified, define time up to the cut off time tau.
   \item Compute the CDF using the proportion of the size in the two arms and \code{npsruvSS::psurv()}.
   \item Return the Fleming-Harriongton weights for weighted Log-rank test.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

